feat(vm-runner): add socket activation with tcp port forwarding
Implement on-demand VM startup via socket activation with TCP port forwarding.
Rename the package from vm-runner to virby-vm-runner and reorganize the codebase
for improved modularity and nix-darwin integration.
