+++ pkgs/vm-runner/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "virby-vm"
+name = "virby-vm-runner"
 version = "0.1.0"
 description = "Vfkit-based VM runner for Virby, with automatic IP discovery and lifecycle management"
 requires-python = ">=3.10"
@@ -7,7 +7,7 @@ authors = [{ name = "Quinn Edenfield", email = "quinn@qeden.dev" }]
 dependencies = ["aiofiles>=24.1.0"]
 
 [project.scripts]
-virby-vm = "vm_runner.cli:cli_main"
+virby-vm = "virby_vm_runner.cli:cli_main"
 
 [dependency-groups]
 dev = [
@@ -22,7 +22,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/vm_runner"]
+packages = ["src/virby_vm_runner"]
 
 [tool.hatch.build.targets.sdist]
 only-include = ["src"]
@@ -38,7 +38,7 @@ use_parentheses = true
 [tool.mypy]
 check_untyped_defs = true
 mypy_path = ["src"]
-packages = ["vm_runner"]
+packages = ["virby_vm_runner"]
 pretty = true
 python_version = "3.13"
 warn_no_return = true
